Rename film repository interface & implementation
