+++ app/src/main/java/com/victormordur/gihbli/app/domain/usecase/flowable/GetUserFilteredFilms.kt
@@ -1,11 +1,11 @@
 package com.victormordur.gihbli.app.domain.usecase.flowable
 
 import com.victormordur.gihbli.app.domain.model.Film
-import com.victormordur.gihbli.app.domain.repository.FilmRepositoryContract
+import com.victormordur.gihbli.app.domain.repository.FilmRepository
 import com.victormordur.gihbli.app.domain.usecase.FlowableUseCase
 import kotlinx.coroutines.flow.mapLatest
 
-abstract class GetUserFilteredFilms(private val repository: FilmRepositoryContract) :
+abstract class GetUserFilteredFilms(private val repository: FilmRepository) :
     FlowableUseCase<List<Film>> {
     override fun requestFlow() =
         repository.getUserFilmsFlow().mapLatest { it.filter { film -> evaluateCondition(film) } }
