+++ app/src/main/java/com/victormordur/gihbli/app/domain/usecase/flowable/GetUserToBeWatchedFilms.kt
@@ -1,9 +1,9 @@
 package com.victormordur.gihbli.app.domain.usecase.flowable
 
 import com.victormordur.gihbli.app.domain.model.Film
-import com.victormordur.gihbli.app.domain.repository.FilmRepositoryContract
+import com.victormordur.gihbli.app.domain.repository.FilmRepository
 
-class GetUserToBeWatchedFilms(repository: FilmRepositoryContract) :
+class GetUserToBeWatchedFilms(repository: FilmRepository) :
     GetUserFilteredFilms(repository) {
     override fun evaluateCondition(film: Film) = !film.watched
 }
